serverutil/web: move Upspin web server into separate package
This package makes Upspin space available through an http.Handler.
